move types to types folder activate lint in ci move tests to test folder
