+++ .taprc
@@ -1,3 +1,6 @@
 100: true
 nyc-arg: ["--exclude=.next"]
 timeout: 120
+
+files:
+  - test/**/*.test.js
