+++ examples/example.js
@@ -3,12 +3,12 @@
 const fastify = require('fastify')()
 
 fastify
-  .register(require('./index'))
+  .register(require('..'))
   .after(() => {
     fastify.next('/hello')
   })
 
-fastify.listen(3000, err => {
+fastify.listen({ port: 3000 }, err => {
   if (err) throw err
   console.log('Server listenging on http://localhost:3000')
 })
